+++ CHANGELOG.md
@@ -68,6 +68,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - `storm-cli extract` will now convert archive file paths to the local systems path separators
+- archive, compression and CLI tests now handle more edge cases and are green
 
 ### Removed